added GET 3.3 and POST 3.4 tasks
