+++ 03-streams/03-file-server-get/index.js
@@ -1,5 +1,7 @@
 const server = require('./server');
 
-server.listen(3000, () => {
+const port = 3000;
+
+server.listen(port, () => {
   console.log('Server is listening on http://localhost:3000');
 });
