+++ 03-streams/04-file-server-post/LimitSizeStream.js
@@ -2,27 +2,28 @@ const stream = require('stream');
 const LimitExceededError = require('./LimitExceededError');
 
 class LimitSizeStream extends stream.Transform {
-  constructor(options) {
-    super(options);
+    constructor(options) {
+        super(options);
 
-    this.limit = options.limit;
-    this.size = 0;
-    this.isObjectMode = !!options.readableObjectMode;
-  }
-
-  _transform(chunk, encoding, callback) {
-    if (this.isObjectMode) {
-      this.size += 1;
-    } else {
-      this.size += chunk.length;
+        this.limit = options.limit;
+        this.size = 0;
+        this.isObjectMode = !!options.readableObjectMode;
     }
 
-    if (this.size > this.limit) {
-      callback(new LimitExceededError());
-    } else {
-      callback(null, chunk);
+    _transform(chunk, encoding, callback) {
+        if (this.isObjectMode) {
+            this.size += 1;
+        } else {
+            this.size += chunk.length;
+        }
+        // console.log(this.size)
+        // console.log(this.limit)
+        if (this.size > this.limit) {
+            callback(new LimitExceededError());
+        } else {
+            callback(null, chunk);
+        }
     }
-  }
 }
 
 module.exports = LimitSizeStream;
