+++ 03-streams/04-file-server-post/README.md
@@ -13,7 +13,7 @@
     
 При любых ошибках сервер должен, по возможности, возвращать ошибку `500`.
 
-Для ограничения размера загружаемого файла можно воспользоваться классом `LimitSizeStream` из 1-ой задачи 
+Для ограничения размера загружаемого файла можно воспользоваться классом `LimitSizeStream` из 2-го 
 модуля.
 
 Для выполнения запросов можно воспользоваться программой `postman`. Проверить правильность
