+++ 03-streams/04-file-server-post/index.js
@@ -2,4 +2,5 @@ const server = require('./server');
 
 server.listen(3000, () => {
   console.log('Server is listening on http://localhost:3000');
+  console.log(server.address().port)
 });
