--- 05-http-server-streams/01-file-server-get/README.md
@@ -1,16 +1,0 @@
-# Файловый сервер - отдача файла
-
-В данной задаче вам необходимо будет реализовать http-сервер, который по запросу пользователя 
-отдавать файл с диска.
-
-- `GET /[filename]` - получание файла из папки `files` (поддерживаются любые типы файлов).
-    - Если файла на диске нет - сервер должен вернуть ошибку `404`.
-    - Вложенные папки не поддерживаются, при запросе вида `/dir1/dir2/filename` - ошибка `400`.
-    - При обрыве соединения необходимо завершить работу стрима.
-
-При любых ошибках сервер должен, по возможности, возвращать ошибку `500`.
-
-Для выполнения запросов можно воспользоваться программой `postman`. Проверить правильность 
-реализации можно также с помощью тестов.
-
-Запуск приложения осуществляется с помощью команды `node index.js`.
