--- 05-http-server-streams/01-file-server-get/index.js
@@ -1,7 +1,0 @@
-const server = require('./server');
-
-const port = 3000;
-
-server.listen(port, () => {
-  console.log('Server is listening on http://localhost:3000');
-});
