--- 05-http-server-streams/01-file-server-get/request.http
@@ -1,10 +1,0 @@
-GET http://localhost:3000/solution.txt
-Content-Type: text/html;
-Cookie: key=value;
-Cookie: key2=value2;
-
-{
-  "hello": "world",
-  "hi": "England",
-  "ahoj": "Prague"
-}
