--- 05-http-server-streams/01-file-server-get/test/fixtures/index.js
@@ -1,1 +1,0 @@
-console.log('hello world');
