--- 05-http-server-streams/02-file-server-post/LimitSizeStream.js
@@ -1,29 +1,0 @@
-const stream = require('stream');
-const LimitExceededError = require('./LimitExceededError');
-
-class LimitSizeStream extends stream.Transform {
-    constructor(options) {
-        super(options);
-
-        this.limit = options.limit;
-        this.size = 0;
-        this.isObjectMode = !!options.readableObjectMode;
-    }
-
-    _transform(chunk, encoding, callback) {
-        if (this.isObjectMode) {
-            this.size += 1;
-        } else {
-            this.size += chunk.length;
-        }
-        // console.log(this.size)
-        // console.log(this.limit)
-        if (this.size > this.limit) {
-            callback(new LimitExceededError());
-        } else {
-            callback(null, chunk);
-        }
-    }
-}
-
-module.exports = LimitSizeStream;
