--- 05-http-server-streams/02-file-server-post/index.js
@@ -1,6 +1,0 @@
-const server = require('./server');
-
-server.listen(3000, () => {
-  console.log('Server is listening on http://localhost:3000');
-  console.log(server.address().port)
-});
