--- 05-http-server-streams/02-file-server-post/request.http
@@ -1,11 +1,0 @@
-POST http://localhost:3000/solution.txt
-Content-Type: text/html;
-Cookie: key=value;
-Cookie: key2=value2;
-
-{
-  "hello": "world",
-  "hi": "England",
-  "ahoj": "Prague",
-  "你好。": "北京"
-}
