--- 05-http-server-streams/02-file-server-post/test/fixtures/index.js
@@ -1,1 +1,0 @@
-console.log('hello world');
